[AIRFLOW-1314] seperated pod operator classes for readability
